# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Система Управления Банковскими Картами - API
  version: 1.0.0
  description: REST API для управления банковскими картами (CRUD, переводы, блокировка, JWT)
paths:
  /api/v1/auth/login:
    get:
      summary: Получение JWT-токена (Вход)
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [USER, ADMIN]
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string

  /api/v1/admin/cards:
    post:
      summary: Создание новой карты (Администратор)
      security:
        - BearerAuth: []
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Карта успешно создана

  /api/v1/user/cards:
    get:
      summary: Просмотр своих карт с пагинацией (Пользователь)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Список карт пользователя получен

  /api/v1/user/cards/transfer:
    post:
      summary: Перевод денег между своими картами (Пользователь)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fromCardId:
                  type: integer
                toCardId:
                  type: integer
                amount:
                  type: number
      responses:
        '200':
          description: Перевод успешно выполнен

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT